Fix multiline syntax in scripts.haml example
Multi-line script blocks in Haml, such as `:ruby`, enable having multiple lines of embedded Ruby code without prepending each line with `- `. This use case is now correctly reflected in the scripts.haml test case.

This also exposes the bug with multi-line syntax highlighting (#12), which will be fixed in a follow-up.